+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 - Added back battery level reporting to crash reporting. Battery level is only reported if battery was enabled before.
 - Added new methods for changing the device id: `changeDeviceIDWithMerge:`, `changeDeviceIDWithoutMerge:`.
 - Deprecated existing method to change the device id: `setNewDeviceID:`
+- Deprecated `attributionID` initial config flag
+- Deprecated `recordAttributionID` method
 
 ## 23.02.2
 - Added server configuration functionality. This is an experimental feature.